Fix small bug with parameter parsing
Prior to this commit we were checking if the size of `platforms`
was greater than zero to see if there were any options passed in
via the --platforms flag. However, when the flag is not used,
`platforms` will be nil meaning that checking the size of it causes
the ci-triage command to error out.

In order to fix this, check if it is nil to determine if we have
options rather than checking the size.