Detect failures in branches besides master
Prior to this commit, we were only checking for failures on the
master branch of a module pipeline. This meant that if master was
green and another branch was red we would get a false positive.

The logic to make this happen is a little complicated because some
modules have a "stable" branch and some have a "release" branch, so
some extra checking is involved to figure out which the module has.

Right now, the output only prints success or failure of a module based
on all branches, but I'd like to add some granulatiry so that the output
displays the status of each branch.