Add --verbose output option
Prior to this commit, ci-triage was taking into account other
branches but there was no way to view an explicit breakdown of each
branch or see the full list of failing jobs.

Add the --verbose option which will print the status of each
individual branch as well as a complete list of failing jobs for
any given branch.